refactor into vmmanager class
