+++ server/vm/utils.ts
@@ -1,0 +1,19 @@
+export const cloudInitWithTls = (
+  host: string,
+  imageName: string
+) => `#!/bin/bash
+until nslookup ${host}
+do
+sleep 5
+echo "Trying DNS lookup again..."
+done
+    
+# Generate cert with letsencrypt
+certbot certonly --standalone -n --agree-tos --email howardzchung@gmail.com --domains ${host}
+chmod -R 755 /etc/letsencrypt/archive
+
+# start browser
+iptables -t nat -A PREROUTING -p tcp --dport 443 -j REDIRECT --to-port 5000
+sed -i 's/scripts-user$/\[scripts-user, always\]/' /etc/cloud/cloud.cfg
+docker run -d --rm --name=vbrowser -v /etc/letsencrypt/archive/${host}:/cert -v /usr/share/fonts:/usr/share/fonts --log-opt max-size=1g --net=host --shm-size=1g --cap-add="SYS_ADMIN" -e DISPLAY=":99.0" -e SCREEN="1280x720@30" -e NEKO_PASSWORD=$(hostname) -e NEKO_PASSWORD_ADMIN=$(hostname) -e NEKO_BIND=":5000" -e NEKO_EPR=":59000-59100" -e NEKO_KEY="/cert/privkey1.pem" -e NEKO_CERT="/cert/fullchain1.pem" ${imageName}
+`;
